+++ apps/runtime-demo/3008-runtime-remote/CHANGELOG.md
@@ -1,5 +1,12 @@
 # 3008-runtime-remote
 
+## 1.0.39
+
+### Patch Changes
+
+- Updated dependencies [70a1708]
+  - @module-federation/enhanced@0.6.9
+
 ## 1.0.38
 
 ### Patch Changes