+++ packages/enhanced/CHANGELOG.md
@@ -1,5 +1,21 @@
 # [0.2.0-canary.5](https://github.com/module-federation/core/compare/enhanced-0.2.0-canary.4...enhanced-0.2.0-canary.5) (2023-11-20)
 
+## 0.6.9
+
+### Patch Changes
+
+- 70a1708: Added a check to skip processing when virtualRuntimeEntry is present.
+
+  - Added an early return in `FederationRuntimePlugin` to skip processing if `options.virtualRuntimeEntry` is defined.
+  - @module-federation/rspack@0.6.9
+  - @module-federation/sdk@0.6.9
+  - @module-federation/runtime-tools@0.6.9
+  - @module-federation/managers@0.6.9
+  - @module-federation/manifest@0.6.9
+  - @module-federation/dts-plugin@0.6.9
+  - @module-federation/bridge-react-webpack-plugin@0.6.9
+  - @module-federation/data-prefetch@0.6.9
+
 ## 0.6.8
 
 ### Patch Changes